test: fold octree_size_field tests into test_octree
